Config.uk: Disable default LIBMUSL_COMPLEX
When using Clang to build Musl with complex numbers support, a linking
error appears: `__muldc3` symbol is undefined. For Clang, the `__muldc3`
symbol is implemented by `libcompiler-rt`.

As complex number support is not typically required by applications
(that use Musl), this commit disable it by setting the `LIBMUSL_COMPLEX`
option to `n`. If required, the application will enable
`LIBMUSL_COMPLEX` (set it to `y`) and, in the case of Clang, add
`libcompiler-rt`.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Andrei Tatar <andrei@unikraft.io>
Approved-by: Maria Sfiraiala <maria.sfiraiala@gmail.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #47