patches: Delete patching of `ioctl` call in `isatty`
Application such as `python3` use `isatty()`, which does an
`ioctl()` with `TIOCGWINSZ` as an argument for checking whether it
should invoke the interactive interpretor or just start interpreting
a given file.

Therefore, we may want to allow such function call to return an error
if called upon a file from a non-pseudo filesystem such as
`ramfs` or `9pfs`.

Furthermore, since this patch is the middle one among the existing
patches from an order standpoint, rename the starting index of the
other patches to reflect the removal.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>