Remove clone patches and use our own defined clone assembly wrappers
We are now using our own `__clone` assembly wrappers in the unikernel
so drop current `__clone` assembly functions that translate libc wrapper
arguments to syscall arguments and instead copy/reuse those defined by
us in the core repository.

Do this copy-paste so that in case if int the future there will appear
differences between what we do in the core in those wrappers and what
we need for musl, the musl wrappers stay the same.

WARNING: There is, however, a very notable difference.

There is a bug in our current native/musl glue code where there is
a mismatch between what we do with TLS size/alignment and what musl
does.

To avoid it (and possibly unknowingly break other things) we have to
keep the forced alignment of the child stack.

This bug has always existed, including in the clone wrapper we are
replacing which also did this alignment. The wrapper claimed something
about `movaps` that is not entirely true. Essentially, because of an
old clone patch (part of the same series that ported musl clone
to Unikraft) that also patches `pthread_create` claiming Unikraft needs
an aligned TLS (for some reason we do not yet know), the clone child
stack ends up misaligned by 8, which causes a `movaps` that somewhere
along the line tries to store something at `some_aligned_offset(%rbp)`
to generate a #GP because rbp itself is not aligned.
To workaround this, back then this was the solution: a forced child
stack alignment. So keep this workaround for now, and debug this ASAP.

As for why we have this alignment for ARM64 as well, I can not be
sure, but I guess it simply aligned itself with the implementation of
x86 since it was introduced afterwards. Besides, it also makes use
of the same pthread_create code. Thus, keep aligning stuff for ARM64
as well.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>